Extract AbstractEmbeddedHedera
